+++ src/Logger/HandlerAwareLogger.php
@@ -21,6 +21,7 @@
 namespace YoutubeDownloader\Logger;
 
 use DateTime;
+use Psr\Log\AbstractLogger;
 use YoutubeDownloader\Logger\Handler\Entry;
 use YoutubeDownloader\Logger\Handler\Handler;
 use YoutubeDownloader\Logger\Handler\SimpleEntry;
@@ -28,7 +29,7 @@ use YoutubeDownloader\Logger\Handler\SimpleEntry;
 /**
  * a logger instance, that works with handler
  */
-class HandlerAwareLogger implements Logger
+class HandlerAwareLogger extends AbstractLogger
 {
     /**
      * @var YoutubeDownloader\Logger\Handler\Handler[]
@@ -45,105 +46,6 @@ class HandlerAwareLogger implements Logger
     public function __construct(Handler $handler)
     {
         $this->addHandler($handler);
-    }
-
-    /**
-     * System is unusable.
-     *
-     * @param string $message
-     * @param array  $context
-     */
-    public function emergency($message, array $context = [])
-    {
-        $this->log(LogLevel::EMERGENCY, $message, $context);
-    }
-
-    /**
-     * Action must be taken immediately.
-     *
-     * Example: Entire website down, database unavailable, etc. This should
-     * trigger the SMS alerts and wake you up.
-     *
-     * @param string $message
-     * @param array  $context
-     */
-    public function alert($message, array $context = [])
-    {
-        return $this->log(LogLevel::ALERT, $message, $context);
-    }
-
-    /**
-     * Critical conditions.
-     *
-     * Example: Application component unavailable, unexpected exception.
-     *
-     * @param string $message
-     * @param array  $context
-     */
-    public function critical($message, array $context = [])
-    {
-        return $this->log(LogLevel::CRITICAL, $message, $context);
-    }
-
-    /**
-     * Runtime errors that do not require immediate action but should typically
-     * be logged and monitored.
-     *
-     * @param string $message
-     * @param array  $context
-     */
-    public function error($message, array $context = [])
-    {
-        return $this->log(LogLevel::ERROR, $message, $context);
-    }
-
-    /**
-     * Exceptional occurrences that are not errors.
-     *
-     * Example: Use of deprecated APIs, poor use of an API, undesirable things
-     * that are not necessarily wrong.
-     *
-     * @param string $message
-     * @param array  $context
-     */
-    public function warning($message, array $context = [])
-    {
-        return $this->log(LogLevel::WARNING, $message, $context);
-    }
-
-    /**
-     * Normal but significant events.
-     *
-     * @param string $message
-     * @param array  $context
-     */
-    public function notice($message, array $context = [])
-    {
-        return $this->log(LogLevel::NOTICE, $message, $context);
-    }
-
-    /**
-     * Interesting events.
-     *
-     * Example: User logs in, SQL logs.
-     *
-     * @param string $message
-     * @param array  $context
-     */
-    public function info($message, array $context = [])
-    {
-        return $this->log(LogLevel::INFO, $message, $context);
-    }
-
-    /**
-     * Detailed debug information.
-     *
-     * @param string $message
-     * @param array  $context
-     */
-    public function debug($message, array $context = [])
-    {
-        return $this->log(LogLevel::DEBUG, $message, $context);
     }
 
     /**